refactor(manager/flux): extract helm repo handling to helper functions
