+++ lib/modules/manager/flux/common.ts
@@ -1,4 +1,6 @@
 import { regEx } from '../../../util/regex';
+import type { HelmRepository } from './schema';
+import type { FluxManifest } from './types';
 
 export const systemManifestFileNameRegex = '(?:^|/)gotk-components\\.ya?ml$';
 
@@ -7,4 +9,20 @@ export const systemManifestHeaderRegex =
 
 export function isSystemManifest(file: string): boolean {
   return regEx(systemManifestFileNameRegex).test(file);
+}
+
+export function collectHelmRepos(manifests: FluxManifest[]): HelmRepository[] {
+  const helmRepositories: HelmRepository[] = [];
+
+  for (const manifest of manifests) {
+    if (manifest.kind === 'resource') {
+      for (const resource of manifest.resources) {
+        if (resource.kind === 'HelmRepository') {
+          helmRepositories.push(resource);
+        }
+      }
+    }
+  }
+
+  return helmRepositories;
 }
